Refactor DTOs y mappers; mejoras en OrderService
- Actualización de DTOs de producto, usuario y orden para reflejar nuevos campos y estructuras.
- Implementación de OrderMapper con mapeo de listas de productos y objetos anidados.
- Ajustes en ProductCatalogMapper y entidades Product/ProductCatalog para soportar el nuevo flujo de datos.
- Actualización de OrderServiceImpl para integración con los mappers y simplificación de la lógica de respuesta.
- Modificación del esquema SQL para reflejar cambios en la estructura de productos y órdenes.

+++ sql/create-schema.sql
@@ -87,7 +87,7 @@ CREATE TABLE products_catalog (
                                   product_name VARCHAR(200) NOT NULL,
                                   description TEXT,
                                   price NUMERIC(10,2) NOT NULL,
-                                  stock_quantity BIGINT DEFAULT 0,
+                                  stock_quantity INTEGER DEFAULT 0,
                                   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
