+++ src/main/java/com/carnerero/agustin/ecommerceapplication/dtos/requests/ProductRequestDTO.java
@@ -20,6 +20,6 @@ public class ProductRequestDTO {
     private UUID productCatalogId;
     @NotNull(message="Quantity is required")
     @Min(value = 1, message = "Quantity must be at least 1")
-    private BigInteger quantity;
+    private Integer quantity;
 
   }
